chore(nossas): refactor import to page select2 field
